Make a ColumnMethodsMixin
