+++ python/cudf/cudf/core/column/categorical.py
@@ -12,11 +12,12 @@ import cudf._lib as libcudf
 from cudf._lib.transform import bools_to_mask
 from cudf.core.buffer import Buffer
 from cudf.core.column import column
+from cudf.core.column.methods import ColumnMethodsMixin
 from cudf.core.dtypes import CategoricalDtype
 from cudf.utils.dtypes import min_signed_type, min_unsigned_type
 
 
-class CategoricalAccessor(object):
+class CategoricalAccessor(ColumnMethodsMixin):
     def __init__(self, column, parent=None):
         """
         Accessor object for categorical properties of the Series values.
@@ -692,25 +693,6 @@ class CategoricalAccessor(object):
             ordered=ordered,
         )
 
-    def _return_or_inplace(self, new_col, **kwargs):
-        """
-        Returns an object of the type of the column owner or updates the column
-        of the owner (Series or Index) to mimic an inplace operation
-        """
-        from cudf.core.index import CategoricalIndex
-
-        owner = self._parent
-        inplace = kwargs.get("inplace", False)
-        if inplace:
-            self._column._mimic_inplace(new_col, inplace=True)
-        else:
-            if owner is None:
-                return new_col
-            elif isinstance(owner, CategoricalIndex):
-                return CategoricalIndex(new_col, name=owner.name)
-            elif isinstance(owner, cudf.Series):
-                return cudf.Series(new_col, index=owner.index, name=owner.name)
-
 
 class CategoricalColumn(column.ColumnBase):
     """Implements operations for Columns of Categorical type
@@ -1206,7 +1188,6 @@ class CategoricalColumn(column.ColumnBase):
         out = super()._mimic_inplace(other_col, inplace=inplace)
         if inplace:
             self._codes = other_col._codes
-
         return out
 
     def view(self, dtype):
